requirements.txt
python-dotenv
python-telegram-bot==13.3
bs4
requests

